+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 ### Added
+- recurse_roles to config
+
 ### Fixed
 ### Changed
 - RdapNotFoundError now inherits from RdapHTTPError instead of LookupError